AFHTTPRequestSerializer: Create default headers only once.
It seems that the headers created in the initializer
`preferredLanguages` and `userAgent` are calculated from values that
don't change throughout the entire program execution. Therefore they can
be created only once. This speeds up the creation of an
`AFHTTPRequestSerializer` which can be significant if for example one
tries to serialize a request in order to get a cached response from
`NSURLCache`.