AFURLRequestSerialization: Remove headers added by default to requests.
We found out that when the app's version changes, cached responses that
are still in the cache are ignored when doing a networking request.
After investigating further the problem stems from Apple's network
library, `CFNetwork`, that compares the user agent HTTP header with the
one of the cached request. If they don't match, a network request is
invoked instead of taking the cached response.

The solution is simply to remove the user agent header that AFNetworking
adds by default.

Furthermore, `CFNetwork` itself sends this header,d this time it doesn't
interfere with the cache.

Before this change:

User-Agent: MyApp/1.5.6 (iPhone; iOS 13.3.1; Scale/2.00)

After this change:

User-Agent: MyApp/1.5.6 CFNetwork/1121.2.2 Darwin/19.3.0