project: support application extensions.
Without the given flag, this raises a warning for
VenusNotificationService that we are linking against a dylib which is
not safe for use in application extensions.
This warning occurred because AFNetworking was added as a dependency
to Intelligence (https://github.com/Lightricks/Foundations/pull/5850),
which VenusNotificationService uses.

Discussed this with The Eshed and he told me that it is safe to allow
this for app extensions.

Verified that after enabling this flag, the warning disappears.