AFSecurityPolicy: avoid using deprecated method.
The method `AFPublicKeyForCertificate` which just extracts the public
key from a certificate used the method `SecTrustEvaluate` to do it.
That method is deprecated and considered "heavy" - every time it's
called from the main thread we get a runtime warning to not do this
and we do this a lot.

This commit makes the method extract the public key in a much simpler
way, using a dedicated method. SSL pinning was tested and seems to be
in working order after this change, and the runtime errors are gone.