AFURLRequestSerialization: Update header dispatch queues.
We have identified an issue related to thread deadlock while
using dispatch_barrier_async queues for certain header
updates. This problem manifests consistently when attempting
to execute copyWithZone concurrently across multiple threads.

The interaction of two specific calls, setValue for the header
and a synchronous call waiting for an update, seems to be the
source of the issue. Under concurrent conditions, these two
functions can deadlock, causing all the relevant threads to
be blocked.

To rectify this issue, we have implemented changes to align
with the behaviour of our upstream branch. This PR was created:
https://github.com/AFNetworking/AFNetworking/pull/4474 that makes
the changes to stop another issue but has been included since
2020 so should have confidence in these updates.